Consolidate card data files
