+++ src/data/index.ts
@@ -4,34 +4,15 @@
  */
 
 import { Card } from '../lib/types';
-import { serviceCards } from './cards';
-import { serviceCards2 } from './cards2';
-import { serviceCards3 } from './cards3';
-import { serviceCards4 } from './cards4';
-import { serviceCards5 } from './cards5';
-import { serviceCards6 } from './cards6';
-import { serviceCards7 } from './cards7';
-import { supportCards } from './supportCards';
-import { supportCards2 } from './supportCards2';
+import { allServiceCards } from './serviceCards';
+import { allSupportCards } from './supportCards';
 import { challenges } from './challenges';
 import { ec2ManagementCards } from './ec2ManagementCards';
 import { ec2ManagementChallenges } from './ec2ManagementChallenges';
 
-// Combine all service cards
-export const allServiceCards: Card[] = [
-  ...serviceCards,
-  ...serviceCards2,
-  ...serviceCards3,
-  ...serviceCards4,
-  ...serviceCards5,
-  ...serviceCards6,
-  ...serviceCards7
-];
-
-// Combine all support cards
-export const allSupportCards: Card[] = [
-  ...supportCards
-];
+// Re-export for convenience
+export { allServiceCards } from './serviceCards';
+export { allSupportCards } from './supportCards';
 
 // All cards combined
 export const allCards: Card[] = [
