+++ src/data/supportCards.ts
@@ -1,11 +1,11 @@
-/**
- * サポートカードデータ
- * サービスカードを補助するサポートカードを定義します
- */
-
 import { Card } from '../lib/types';
 
-export const supportCards: Card[] = [
+/**
+ * サポートカードデータ
+ * すべてのサポートカードをまとめた単一のファイル
+ */
+
+export const allSupportCards: Card[] = [
   {
     id: 101,
     name: 'FinOps',
@@ -185,7 +185,141 @@ export const supportCards: Card[] = [
     effect: 'セキュリティ関連のカードと組み合わせると、状況レベルが30%向上します。また、セキュリティ関連の課題では、コストが25%削減されます。',
     description: '「信頼しない、常に検証する」の原則に基づき、ネットワーク内外の脅威からシステムを保護します。最小権限の原則を徹底します。',
     keywords: ['セキュリティ', '認証', 'アクセス制御']
+  },
+  {
+    id: 116,
+    name: '多要素認証',
+    type: 'support',
+    cost: 0,
+    effect: 'IAMユーザーのセキュリティを強化し、不正アクセスを防止します。',
+    description: 'ユーザー名とパスワードに加えて、追加の認証要素を要求します。',
+    keywords: ['セキュリティ', '認証', 'アクセス制御']
+  },
+  {
+    id: 117,
+    name: 'データ暗号化',
+    type: 'support',
+    cost: 1,
+    effect: '保存データと転送中のデータを暗号化し、セキュリティを強化します。',
+    description: 'データを暗号化することで、不正アクセスからデータを保護します。',
+    keywords: ['セキュリティ', '暗号化', 'データ保護']
+  },
+  {
+    id: 118,
+    name: 'ディザスタリカバリ計画',
+    type: 'support',
+    cost: 2,
+    effect: '災害時にもサービスを継続できるよう、バックアップと復旧手順を確立します。',
+    description: '災害発生時にビジネス継続性を確保するための計画です。',
+    keywords: ['DR', 'バックアップ', '事業継続']
+  },
+  {
+    id: 119,
+    name: 'クロスリージョンレプリケーション',
+    type: 'support',
+    cost: 2,
+    effect: 'データを複数のリージョンに複製し、リージョン障害に対する耐性を高めます。',
+    description: 'データを異なるAWSリージョンに自動的に複製します。',
+    keywords: ['冗長性', 'グローバル', '高可用性']
+  },
+  {
+    id: 120,
+    name: 'キャッシング戦略',
+    type: 'support',
+    cost: 1,
+    effect: 'データベースの負荷を軽減し、アプリケーションのパフォーマンスを向上させます。',
+    description: '頻繁にアクセスされるデータをキャッシュに保存し、高速にアクセスできるようにします。',
+    keywords: ['パフォーマンス', 'キャッシュ', '高速化']
+  },
+  {
+    id: 121,
+    name: 'コンテンツ配信最適化',
+    type: 'support',
+    cost: 1,
+    effect: 'グローバルユーザーへのコンテンツ配信を高速化します。',
+    description: 'エッジロケーションを活用して、ユーザーに近い場所からコンテンツを配信します。',
+    keywords: ['CDN', 'パフォーマンス', 'グローバル']
+  },
+  {
+    id: 122,
+    name: 'サーバーサイド暗号化',
+    type: 'support',
+    cost: 0,
+    effect: 'データをサーバー側で自動的に暗号化し、セキュリティを強化します。',
+    description: 'データがディスクに書き込まれる前に暗号化され、読み取り時に復号化されます。',
+    keywords: ['セキュリティ', '暗号化', 'データ保護']
+  },
+  {
+    id: 123,
+    name: 'ウェルアーキテクテッド・フレームワーク',
+    type: 'support',
+    cost: 0,
+    effect: 'AWSのベストプラクティスに基づいてアーキテクチャを設計し、品質を向上させます。',
+    description: '安全で高性能、回復力のある、効率的なインフラストラクチャを構築するためのフレームワークです。',
+    keywords: ['ベストプラクティス', 'アーキテクチャ', '品質']
+  },
+  {
+    id: 124,
+    name: 'クラウドネイティブ設計',
+    type: 'support',
+    cost: 1,
+    effect: 'クラウドの利点を最大限に活用したアプリケーション設計を行います。',
+    description: 'クラウド環境の特性を活かした設計手法で、スケーラビリティや弾力性を向上させます。',
+    keywords: ['クラウドネイティブ', '設計', '最適化']
+  },
+  {
+    id: 125,
+    name: 'DevOpsプラクティス',
+    type: 'support',
+    cost: 1,
+    effect: '開発と運用の連携を強化し、デプロイの頻度と品質を向上させます。',
+    description: '開発と運用のプロセスを統合し、ソフトウェア開発のライフサイクルを加速します。',
+    keywords: ['DevOps', '自動化', '継続的デリバリー']
+  },
+  {
+    id: 126,
+    name: 'インフラストラクチャのモニタリング',
+    type: 'support',
+    cost: 1,
+    effect: 'システムの健全性とパフォーマンスを継続的に監視し、問題を早期に検出します。',
+    description: 'メトリクスの収集、分析、アラートを通じて、システムの状態を可視化します。',
+    keywords: ['モニタリング', '可視化', 'アラート']
+  },
+  {
+    id: 127,
+    name: 'ログ集約',
+    type: 'support',
+    cost: 1,
+    effect: '複数のサービスからログを収集し、分析と問題解決を容易にします。',
+    description: '分散システム全体からログを一元的に収集し、検索可能にします。',
+    keywords: ['ログ', '分析', 'トラブルシューティング']
+  },
+  {
+    id: 128,
+    name: 'サービスメッシュ',
+    type: 'support',
+    cost: 2,
+    effect: 'マイクロサービス間の通信を管理し、可視性と制御を向上させます。',
+    description: 'サービス間の通信を処理するための専用インフラストラクチャレイヤーです。',
+    keywords: ['マイクロサービス', '通信', '可視性']
+  },
+  {
+    id: 129,
+    name: 'APIファースト設計',
+    type: 'support',
+    cost: 1,
+    effect: 'APIを中心にアプリケーションを設計し、拡張性と再利用性を向上させます。',
+    description: 'APIを最初に設計し、それを中心にアプリケーションを構築する手法です。',
+    keywords: ['API', '設計', '拡張性']
+  },
+  {
+    id: 130,
+    name: 'クラウドセキュリティアセスメント',
+    type: 'support',
+    cost: 1,
+    effect: 'セキュリティの脆弱性を特定し、対策を講じることでセキュリティを強化します。',
+    description: 'クラウド環境のセキュリティ状態を評価し、改善点を特定します。',
+    keywords: ['セキュリティ', '評価', 'コンプライアンス']
   }
 ];
 
-export const supportCards2: Card[] = [];
